colocate the DictationProvider to the open ai dictation route
